requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2==2.9.9
drf-spectacular==0.26.5